Add game logic
