+++ bin/main.rb
@@ -1,74 +1,6 @@
 #!/usr/bin/env ruby
 require 'colorize'
-# require '../lib/TicTacToe.rb'
-
-class Game
-  # Class to be implemented inside TicTacToe Module
-  attr_reader :board, :victory
-
-  def initialize
-    @board = (1..9).to_a
-    @victory = Array.new(9, false)
-  end
-
-  def mark?(number, symbol)
-    return false if number < 1 || number > 9 || number != @board[number - 1]
-
-    @board[number - 1] = symbol
-    true
-  end
-
-  def check_for_rows?
-    (1..9).step(3).each do |i|
-      if @board[i - 1] == @board[i] && @board[i] == @board[i + 1]
-        return @victory[i - 1] = @victory[i] = @victory[i + 1] = true
-      end
-    end
-    false
-  end
-
-  def check_for_columns?
-    (3..5).each do |i|
-      if @board[i - 3] == @board[i] && @board[i] == @board[i + 3]
-        return @victory[i - 3] = @victory[i] = @victory[i + 3] = true
-      end
-    end
-    false
-  end
-
-  def check_for_main_diagonal?
-    @victory[0] = @victory[4] = @victory[8] = true if @board[0] == @board[4] && @board[4] == @board[8]
-  end
-
-  def check_for_secondary_diagonal?
-    @victory[2] = @victory[4] = @victory[6] = true if @board[2] == @board[4] && @board[4] == @board[6]
-  end
-
-  def winner?
-    result = false
-    result ||= check_for_rows?
-    result ||= check_for_columns?
-    result ||= check_for_main_diagonal?
-    result ||= check_for_secondary_diagonal?
-    result
-  end
-
-  def cats_game?
-    @board.all?(String)
-  end
-end
-
-class Players
-  # To be implemented inside TicTacToe Module
-  attr_reader :match_players
-
-  def initialize(name1, name2)
-    @match_players = [
-      { name: name1, symbol: 'X' },
-      { name: name2, symbol: 'O' }
-    ]
-  end
-end
+require_relative '../lib/tic_tac_toe'
 
 def clear_screen
   system('cls') || system('clear')
